Move gRPC transport code into new `transport/grpcext` package
Extracts `NewClient`, `Option`, `WithDefaultTimeout`, the interceptors,
and the `auth` type from `internal/deviceutil` into a dedicated
`internal/transport/grpcext` package, mirroring the existing
`internal/transport/gnmiext` sibling. `deviceutil` now only contains
device CRD helpers and the `Connection` type.
